requirements.txt
google-generativeai==0.3.2
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
python-telegram-bot==20.7
